+++ apps/mobile/src/app/(tabs)/(home,search,profile,notifications,settings)/settings/defaults.tsx
@@ -204,7 +204,7 @@ export default function Screen() {
             if (item.type === 'feed') {
               update({
                 community: undefined,
-                feed: item.feed.name,
+                feed: item.feed.id,
                 feedType: undefined,
               })
             }