+++ apps/mobile/src/transformers/feed.ts
@@ -1,5 +1,6 @@
 import { fromUnixTime } from 'date-fns'
 import { decode } from 'entities'
+import { compact } from 'lodash'
 
 import { type FeedDataSchema } from '~/schemas/feeds'
 import { type Feed } from '~/types/feed'
@@ -8,8 +9,12 @@ export function transformFeed(data: FeedDataSchema): Feed {
   return {
     communities: data.subreddits.map((community) => community.name),
     createdAt: fromUnixTime(data.created_utc),
-    id: data.name,
+    id: getId(data.path) ?? data.name,
     image: data.icon_url ? decode(data.icon_url) || undefined : undefined,
     name: data.display_name,
   }
+}
+
+function getId(path: string) {
+  return compact(path.split('/')).pop()
 }